svn:revision:77321
svn:branch:thirdparty_rework
svn:account:starseeker

+++ bench/pixcmp.c
@@ -215,15 +215,15 @@ main(int argc, char *argv[])
     if ((sf1.st_size - f1_skip) != (sf2.st_size - f2_skip)) {
 	bu_log("WARNING: Different image sizes detected\n");
 	if (print_bytes) {
-	    bu_log("\t%s: %7zu bytes (%8zu bytes, skipping %7zu)\n",
-		   argv[0], sf1.st_size - f1_skip, sf1.st_size, f1_skip);
-	    bu_log("\t%s: %7zu bytes (%8zu bytes, skipping %7zu)\n",
-		   argv[1], sf2.st_size - f2_skip, sf2.st_size, f2_skip);
+	    bu_log("\t%s: %7llu bytes (%8llu bytes, skipping %7llu)\n",
+		   argv[0], (unsigned long long)(sf1.st_size - f1_skip), (unsigned long long)sf1.st_size, (unsigned long long)f1_skip);
+	    bu_log("\t%s: %7llu bytes (%8llu bytes, skipping %7llu)\n",
+		   argv[1], (unsigned long long)(sf2.st_size - f2_skip), (unsigned long long)sf2.st_size, (unsigned long long)f2_skip);
 	} else {
-	    bu_log("\t%s: %7zu pixels (%8zu bytes, skipping %7zu)\n",
-		   argv[0], (sf1.st_size - f1_skip)/3, sf1.st_size, f1_skip);
-	    bu_log("\t%s: %7zu pixels (%8zu bytes, skipping %7zu)\n",
-		   argv[1], (sf2.st_size - f2_skip)/3, sf2.st_size, f2_skip);
+	    bu_log("\t%s: %7llu pixels (%8llu bytes, skipping %7llu)\n",
+		   argv[0], (unsigned long long)((sf1.st_size - f1_skip)/3), (unsigned long long)sf1.st_size, (unsigned long long)f1_skip);
+	    bu_log("\t%s: %7llu pixels (%8llu bytes, skipping %7llu)\n",
+		   argv[1], (unsigned long long)((sf2.st_size - f2_skip)/3), (unsigned long long)sf2.st_size, (unsigned long long)f2_skip);
 	}
 	bu_exit(1, "ERROR: Cannot pixcmp due to different images sizes (unimplemented).\n");
     }
